test(pascal): the side-effect applier's differential, and every progress log line pinned
New `pascal_side_effects_equivalence`: 49 corpus rows driving v4's real
`applyCustomToolEffects` with `getRepositories` and `writeGeneralState` mocked to
recorders, diffed against `side_effects`' pure planning pass through a new
feature-gated seam. Four comparands per row — the one character write and its
COUNT, the state writes in v4's commit order, the returned `AppliedEffect[]`, and
the warn lines rebuilt whole. Both sides read one committed corpus, so a case
cannot be added to the oracle alone.

Rows 1-15 are v4's pre-progressions state/metadata matrix and rows 16-49 its
progress suite, in that order deliberately: with the progress branch neutered
the first fifteen pass and row sixteen is the first to fail, which is the
order's "green on the existing arms, red on the progress arms" measured rather
than asserted.

Two rows are not v4's. A mutation swapping `value !== true` for JS truthiness on
the `remove` pseudo-field SURVIVED, because v4 tests `remove` with `true` and
`false` only and both rules agree on those; `remove` written as a truthy number
and as a truthy string are the rows that tell them apart. v4 declines both.

Item 6's log lines pinned with presence and silence: the templated
`Custom tool ${subject} test did not match` in both spellings, the two new
`render_template` arms, and `Custom tool progress effect folded`. The fold pin
caught a real gap — v5 rendered `previous=Some(String("..."))` where v4 logs the
bare value and omits the key when it is `undefined`.

Tier 2 measured, neither needing an edit: `chat_run_tool.rs` is not an entrance
(it reaches the applier through `executor.rs`, already threaded), and
`CoreResponse::CustomToolPreview` is an untyped `Value`, so no `api/types.rs`
fence.

Also three things the previous commit's gate missed, having been run against the
wrong worktree's logs: nine clippy findings, a stale in-crate assertion carrying
the pre-widening effect-target refusal sentence, and the `run_custom` catalog
entry in the GENERATED `tools/definitions/data.rs` — regenerated at the pin, one
line moved, the other 57 definitions byte-identical. Seven comments annotating
`1_788_004_800_000` as 2026-09-08T12:00:00Z corrected to 2026-08-29T12:00:00Z.

(cherry picked from commit 2291cfae205798694de236083c4a88a0a572219c)

+++ crates/quilltap-core/src/api/engine.rs
@@ -4749,7 +4749,7 @@ impl CoreEngine {
                     SINGLE_USER_ID,
                     &chat_id,
                     // P4.D169: production's clock — one reading for the listing.
-                    crate::clock::now_unix_ms() as i64,
+                    crate::clock::now_unix_ms(),
                 ),
                 Err(r) => r,
             },
@@ -4777,7 +4777,7 @@ impl CoreEngine {
                         // templates read, `{{now}}`, and the `updatedAt` an effect
                         // stamps — so a tool cannot see one moment and record
                         // another.
-                        crate::clock::now_unix_ms() as i64,
+                        crate::clock::now_unix_ms(),
                     )
                     .await
                 }
@@ -4815,7 +4815,7 @@ impl CoreEngine {
                         // arms never touch it.
                         consult.as_deref(),
                         // P4.D169: the bench's one reading (v4 `benchNowMs`).
-                        crate::clock::now_unix_ms() as i64,
+                        crate::clock::now_unix_ms(),
                     )
                     .await
                 }
@@ -4837,7 +4837,7 @@ impl CoreEngine {
                     llm.as_ref(),
                     // P4.D169: the audit's one reading — the sheet is derived
                     // once and held fixed across every draw.
-                    crate::clock::now_unix_ms() as i64,
+                    crate::clock::now_unix_ms(),
                 ),
                 Err(r) => r,
             },
